# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Yunzai WebUI API
  version: "1.0.0"
  description: |-
    TRSS-Yunzai 的运维面板接口。v1 是**只读**的：所有接口都是 `GET`。

    ## 启用
    在 `config/config/server.yaml` 里设置 `server.webui.enable: true`，且 `server.auth`
    必须非空——否则**不挂载** `/dashboard` 与 `/api/v1/*`（启动日志会给出可照抄的配置）。
    默认是关闭的，且旧配置里没有这个键，所以升级后不会自己打开。

    ## 鉴权
    与宿主既有路由共用 `serverAuth`：`server.auth` 是「请求头名 → 令牌」的映射，
    **请求头或查询参数任一键匹配**即通过，否则 401。头名由配置决定，
    下面的 `Authorization` 只是一个默认示例。

    ## 限流与体积
    `/api/v1/*` 统一挂一个 per-IP 令牌桶（默认突发 60 次、每秒回填 10 个）与 1MB 请求体上限。

    ## 错误形状
    统一为 `{ code, message }`。唯一的例外是 **401**：它由宿主的 `serverAuth` 产生，
    返回的是纯文本（`Unauthorized`），且会同时写入一条访问日志。

    ## 契约与实现的一致性
    `tests/unit/web/openapi.test.js` 会对照本文件与真实路由/真实响应做检查：
    路径与方法必须双边一致，`200` 响应的顶层字段必须与实际返回的字段一致。
    改接口时**必须同步改这里**，否则测试会红。
servers:
  - url: http://localhost:2536
    description: 默认端口，见 `server.yaml` 的 `server.port`
...
security:
  - serverAuth: []
paths:
...
  /api/v1/plugins:
    get:
      tags: [状态]
      summary: 已加载的插件
      description: |-
        数据来自内存里排好序的 `priority`，**不是**磁盘上的 `plugin.json`：
        面板要回答的是"现在生效的是哪些"，而插件被 `disable` 后 `plugin.json` 仍在磁盘上。

        数组顺序就是**执行顺序**（按 `priority` 升序），保持原样不做排序。

        `enabled` 反映的是 `group.yaml` 的 `default.disable` 名单。注意它是
        **按插件名**匹配的，且**更具体的群段可以盖过它**——所以 `enabled: true`
        只表示"没在全局停用名单里"，不代表它在每个群里都真的生效。
      operationId: plugins
      responses:
        "200":
          description: 插件列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plugins"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/logs:
    get:
      tags: [日志]
      summary: 实时日志流（SSE）
      description: |-
        `text/event-stream`。连上后先**回放最近若干条**（默认 200，可用 `?limit=` 调整），
        之后实时推送；每 15 秒发一个注释行 `: ping` 保活。

        每个事件形如：

        ```
        event: log
        data: {"time":1758700000000,"level":"MARK","category":"command","message":"…"}
        ```

        - `level`：log4js 的等级（`TRACE`/`DEBUG`/`INFO`/`WARN`/`ERROR`/`FATAL`/`MARK`）
        - `category`：发日志的 logger 名（`message` 是默认 logger 的名字）
        - `message`：已去掉 ANSI 颜色码

        选 SSE 而不是 WebSocket 的理由见 `docs/refactor/06-webui.md` §3.5：单向、自动重连、
        不需要额外握手。日志来源是给 log4js 追加的 appender，与终端是**同一份**事件
        （落盘文件里没有 `info`/`debug`，那是选择 appender 而不是 tail 文件的原因）。

        ## 客户端注意
        响应头带 `Cache-Control: no-cache, no-transform`——宿主的 `compression()`
        会把小于阈值的分片攒在缓冲区里，而 SSE 全是小分片。
        中间有反向代理时，链路要关掉对 `text/event-stream` 的缓冲。
      operationId: logs
      parameters:
        - name: limit
          in: query
          required: false
          description: 回放条数。非法值或超过上限时退回上限。
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 200
      responses:
        "200":
          description: SSE 流（长连接；断开时服务端释放订阅）
          content:
            text/event-stream:
              schema:
                type: string
                description: 每帧是一个 log 事件（data 行里是 JSON），或一行保活注释
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
components:
  securitySchemes:
    serverAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |-
        头名与令牌都来自 `server.auth` 配置（键名 = 请求头名）。这里的 `Authorization`
        只是默认示例；查询参数同名匹配也可以。
  responses:
    Unauthorized:
      description: |-
        鉴权失败。由宿主的 `serverAuth` 产生，响应体是**纯文本** `Unauthorized`，
        同时会写一条 `HTTP … 鉴权失败` 的访问日志。
      content:
        text/plain:
          schema:
            type: string
            examples: [Unauthorized]
    RateLimited:
      description: 超过 per-IP 令牌桶。`Retry-After` 是建议等待的秒数。
      headers:
        Retry-After:
          schema:
            type: integer
          description: 建议等待秒数
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |-
            稳定的机器可读码：`bad_request` / `not_found` / `rate_limited` /
            `payload_too_large` / `length_required` / `starting` / `internal_error`
          examples: [bad_request]
        message:
          type: string
          description: 给人看的说明（会随版本变化，不要用它做判断）
      additionalProperties: false
...
    Plugins:
      type: object
      required: [total, plugins]
      properties:
        total:
          type: integer
        plugins:
          type: array
          items:
            $ref: "#/components/schemas/Plugin"
    Plugin:
      type: object
      required: [name, namespace, key, priority, description, events, rules, enabled]
      properties:
        name:
          type: [string, "null"]
          description: 插件展示名。**启停就是按它匹配的**（`group.yaml` 的 `disable` 名单）
        namespace:
          type: [string, "null"]
        key:
          type: [string, "null"]
          description: 插件文件与类的标识，与日志里出现的那个一致
        priority:
          type: [integer, "null"]
          description: 声明值；为 null 表示没声明（实际顺序见数组下标）
        description:
          type: [string, "null"]
        events:
          type: array
          items:
            type: string
        rules:
          type: array
          items:
            $ref: "#/components/schemas/PluginRule"
        enabled:
          type: [boolean, "null"]
          description: |-
            是否在全局停用名单里。`null` 表示**这个条目没有名字**（`name` 为 null），
            也就无法用启停接口表达——界面应把它的开关禁用。

            注意 `true` 只表示"没被全局停用"，不代表它在每个群里都生效：
            更具体的群段（`enable`）可以盖过全局停用。
...
    PluginRule:
      type: object
      required: [reg, fnc, permission, log]
      properties:
        reg:
          type: string
          description: 正则的 `source`（不带斜杠与标志）
        fnc:
          type: [string, "null"]
          description: 处理方法名
        permission:
          type: [string, "null"]
        log:
          type: [boolean, "null"]
          description: 值为 false 表示该规则的日志降级为 debug；null 表示没声明
...